+++ Magna-World/src/main/java/de/morigm/magnaworld/loader/PluginLoader.java
@@ -2,6 +2,7 @@ package de.morigm.magnaworld.loader;
 
 import de.morigm.magnaworld.Main;
 import de.morigm.magnaworld.commands.MWCREATE;
+import de.morigm.magnaworld.commands.MWTP;
 
 public class PluginLoader 
 {
@@ -9,6 +10,7 @@ public class PluginLoader
 	public void loadCommands()
 	{
 		new MWCREATE().register("mwcreate", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
+		new MWTP().register("mwtp", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
 	}
 	
 	public void loadListener()